Fix: Remove usage of webpack raw loader for Rusha hashing
Don't use webpack raw loader since that won't work when Uploader is imported as a React component.
