+++ .eslintignore
@@ -1,4 +1,3 @@
 flow/**
 lib/**
 dist/**
-src/third-party/**
