+++ package.json
@@ -152,7 +152,6 @@
     "prettier-eslint-cli": "^4.1.1",
     "prop-types": "^15.6.0",
     "properties-parser": "^0.3.1",
-    "raw-loader": "^0.5.1",
     "react": "^15.6.1",
     "react-addons-test-utils": "^15.6.0",
     "react-dom": "^15.6.1",
